This commit merges the initial Babelfish 1.0 commit onto community 14.2.

+++ src/backend/tcop/utility.c
@@ -59,6 +59,7 @@
 #include "commands/vacuum.h"
 #include "commands/view.h"
 #include "miscadmin.h"
+#include "parser/parser.h"
 #include "parser/parse_utilcmd.h"
 #include "postmaster/bgwriter.h"
 #include "rewrite/rewriteDefine.h"
@@ -769,7 +770,12 @@ standard_ProcessUtility(PlannedStmt *pstmt,
 
 		case T_CreatedbStmt:
 			/* no event triggers for global objects */
-			PreventInTransactionBlock(isTopLevel, "CREATE DATABASE");
+			/* In case of TSQL mode, we allow create database from
+			 * transaction blocks to turn batch mode ON by default.
+			 */
+			if (sql_dialect != SQL_DIALECT_TSQL) {
+				PreventInTransactionBlock(isTopLevel, "CREATE DATABASE");
+			}
 			createdb(pstate, (CreatedbStmt *) parsetree);
 			break;
 
@@ -785,7 +791,12 @@ standard_ProcessUtility(PlannedStmt *pstmt,
 
 		case T_DropdbStmt:
 			/* no event triggers for global objects */
-			PreventInTransactionBlock(isTopLevel, "DROP DATABASE");
+			/* In case of TSQL mode, we allow drop database from
+			 * transaction blocks to turn batch mode ON by default.
+			 */
+			if (sql_dialect != SQL_DIALECT_TSQL) {
+				PreventInTransactionBlock(isTopLevel, "DROP DATABASE");
+			}
 			DropDatabase(pstate, (DropdbStmt *) parsetree);
 			break;
